[MI-2230] Update record types for subscription
